% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_func.R
\name{bh.func}
\alias{bh.func}
\title{Benjamini-Hochberg Correction in LAWS-HiC}
\usage{
bh.func(pv, q)
}
\arguments{
\item{pv}{A numeric vector of p-values}

\item{q}{The false discovery rate (FDR) threshold.}
}
\value{
A list with the number of rejected hypotheses, threshold, and decision rule.
}
\description{
Applies the BH procedure for input p-values
}
\examples{
p_vector <- c(0.01, 0.02, 0.03, 0.04, 0.00001, 0.002, 0.003, 0.6)
bh.func(p_vector, 0.005)
}
