% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laws_hic.R
\name{laws_hic}
\alias{laws_hic}
\title{LAWS-Adjusted 3D Peak Calling P-Values}
\usage{
laws_hic(input, domain_input, chr, resolution)
}
\arguments{
\item{input}{A data.table containing Hi-C interactions with columns: \code{chr}, \code{fragmentMid1}, \code{fragmentMid2}, and \code{p_value}.}

\item{domain_input}{A data.table containing TAD boundary information with columns: \code{chr}, \code{x1}, and \code{x2}.}

\item{chr}{A character string or interger indicating the chromosome to process, please make sure it is consistent with your input.}

\item{resolution}{A numeric value representing the resolution of Hi-C data in kilobases (kb).}
}
\value{
A \code{data.table} containing the original peak calling results with two additional columns: \code{p_laws} (adjusted p-values) and \code{pi} (estimated sparsity level).
}
\description{
This function take peak caller results and topologically associating domains (TADs) list as input, adjusting the p-values by LAWS procedure then returns the LAWS-adjusted p-values and the estimated sparsity level.
}
