% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laws_hic_paral.R
\name{laws_hic_paral}
\alias{laws_hic_paral}
\title{LAWS-Adjusted 3D Peak Calling P-Values. (Parallelized)}
\usage{
laws_hic_paral(input, domain_input, chr, resolution)
}
\arguments{
\item{input}{A data.table containing Hi-C interactions with columns:
\code{chr}, \code{fragmentMid1}, \code{fragmentMid2}, and \code{p_value}.}

\item{domain_input}{A data.table containing TAD boundary information with columns:
\code{chr}, \code{x1}, and \code{x2}.}

\item{chr}{A character string or interger indicating the chromosome to process, please make sure it is consistent with your input.}

\item{resolution}{A numeric value representing the resolution of Hi-C data in kilobases (kb).}

\item{gc_frequency}{Garbage collection frequency, default is 20.}
}
\value{
A \code{data.table} containing the original peak calling results with two additional columns: \code{p_laws} (adjusted p-values) and \code{pi} (estimated sparsity level).
}
\description{
This is the Parallelized version of \code{laws_hic}
}
